Adding some extra padding to the screen
Initially there was an issue with the screen. When the program launches it did not show the entire screen. I had to manually resize the window.
Then I figured out that this is happening because Win32 API does not provide the given screen resolution to the canvas. This is rather set as the resolution of the entire window including the title bar and the border. 
After realizing that it was a simple fix. I added extra padding to both width and height